+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# Unreleased
+
+_released `2026-09-15`_
+
+- add `--date` flag to `hark new` (optional, defaults to today)
+- validate that slugs are alphanumeric and/or hyphens
+
 ## 0.1.0
 
 _released `2026-09-14`_